authorization

+++ node_tut/routes/articles.js
@@ -4,10 +4,11 @@ const router = express.Router();
 
 //bring in article model
 let Article = require('../models/article');
+let User = require('../models/user');
 
 
 //add article route
-router.get('/add', (req, res) => {
+router.get('/add', ensureAuthenticated, (req, res) => {
 	res.render('add_article', {
 		title: 'Add article'
 	});
@@ -17,7 +18,7 @@ router.get('/add', (req, res) => {
 router.post('/add', (req, res) => {
 	let article = new Article();
 	article.title = req.body.title;
-	article.author = req.body.author;
+	article.author = req.user._id;
 	article.content = req.body.content;
 	article.isPublic = req.body.isPublic == undefined ? false : true;
 
@@ -32,8 +33,12 @@ router.post('/add', (req, res) => {
 });
 
 //load edit form
-router.get('/edit/:id', (req, res) => {
+router.get('/edit/:id', ensureAuthenticated, (req, res) => {
 	Article.findById(req.params.id, (err, article) => {
+		if(article.author != req.user._id){
+			req.flash('danger', 'Not authorized');
+			return res.redirect('/');
+		}
 		res.render('edit_article', {
 			title: 'Edit article',
 			article: article
@@ -63,23 +68,47 @@ router.post('/edit/:id', (req, res) => {
 
 //delete article
 router.delete('/:id', (req, res) => {
+	if(!req.user._id){
+		res.status(500).send();
+	}
+
 	let query = {_id:req.params.id}
 
-	Article.deleteOne(query, (err) => {
-		if(err){
-			console.log(err);
+	Article.findById(req.params.id, (err, article) => {
+		if(article.author != req.user._id){
+			res.status(500).send();
+		} else {
+			Article.deleteOne(query, (err) => {
+				if(err){
+					console.log(err);
+				}
+				res.send('Success');
+			});
 		}
-		res.send('Success');
 	});
+
 });
 
 //get single article
 router.get('/:id', (req, res) => {
 	Article.findById(req.params.id, (err, article) => {
-		res.render('article', {
-			article: article
+		User.findById(article.author, (err, user) => {
+			res.render('article', {
+				article: article,
+				author: user.name
+			});
 		});
 	});
 });
+
+//access control
+function ensureAuthenticated(req, res, next){
+	if(req.isAuthenticated()){
+		return next();
+	} else {
+		req.flash('danger', 'Please log in');
+		res.redirect('/users/login');
+	}
+}
 
 module.exports = router;
